refactor: Admin Controller
