+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/presentation/response/report/SpotReportAllAdminResponse.kt
@@ -5,14 +5,14 @@ import com.sseudam.support.page.Page
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "사용자 신고 내역 응답 Json")
-data class SpotReportAllResponse(
-    val list: List<SpotReportResponse>,
+data class SpotReportAllAdminResponse(
+    val list: List<SpotReportAdminResponse>,
     val totalCount: Long,
 ) {
     companion object {
-        fun of(page: Page<SpotReport.Detail>): SpotReportAllResponse =
-            SpotReportAllResponse(
-                list = page.content.map { SpotReportResponse.of(it) },
+        fun of(page: Page<SpotReport.Detail>): SpotReportAllAdminResponse =
+            SpotReportAllAdminResponse(
+                list = page.content.map { SpotReportAdminResponse.of(it) },
                 totalCount = page.totalCount,
             )
     }
