+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/presentation/response/suggestion/SpotSuggestionAllAdminResponse.kt
@@ -3,14 +3,14 @@ package com.sseudam.admin.presentation.response.suggestion
 import com.sseudam.suggestion.SpotSuggestion
 import com.sseudam.support.page.Page
 
-data class SpotSuggestionAllResponse(
-    val list: List<SpotSuggestionResponse>,
+data class SpotSuggestionAllAdminResponse(
+    val list: List<SpotSuggestionAdminResponse>,
     val totalCount: Long,
 ) {
     companion object {
         fun of(page: Page<SpotSuggestion.Info>) =
-            SpotSuggestionAllResponse(
-                list = page.content.map { SpotSuggestionResponse.of(it) },
+            SpotSuggestionAllAdminResponse(
+                list = page.content.map { SpotSuggestionAdminResponse.of(it) },
                 totalCount = page.totalCount,
             )
     }
