+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/presentation/response/user/UserAdminResponse.kt
@@ -3,15 +3,15 @@ package com.sseudam.admin.presentation.response.user
 import com.sseudam.user.UserProfile
 import java.time.LocalDateTime
 
-data class UserResponse(
+data class UserAdminResponse(
     val userId: Long,
     val email: String,
     val nickname: String,
     val createdAt: LocalDateTime,
 ) {
     companion object {
-        fun of(userProfile: UserProfile): UserResponse =
-            UserResponse(
+        fun of(userProfile: UserProfile): UserAdminResponse =
+            UserAdminResponse(
                 userProfile.id,
                 userProfile.email,
                 userProfile.nickname,
