+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/presentation/response/user/UserPageAdminResponse.kt
@@ -3,14 +3,14 @@ package com.sseudam.admin.presentation.response.user
 import com.sseudam.support.page.Page
 import com.sseudam.user.UserProfile
 
-data class UserPageResponse(
-    val list: List<UserResponse>,
+data class UserPageAdminResponse(
+    val list: List<UserAdminResponse>,
     val totalCount: Long,
 ) {
     companion object {
         fun of(page: Page<UserProfile>) =
-            UserPageResponse(
-                list = page.content.map { UserResponse.of(it) },
+            UserPageAdminResponse(
+                list = page.content.map { UserAdminResponse.of(it) },
                 totalCount = page.totalCount,
             )
     }
