+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/presentation/v1/annotation/ApiAdminV1Controller.kt
@@ -1,0 +1,11 @@
+package com.sseudam.admin.presentation.v1.annotation
+import org.springframework.stereotype.Component
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@Component
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+@RestController
+@RequestMapping("/api/v1/admin")
+annotation class ApiAdminV1Controller
